requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.4.0
argparse==1.2.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-guardian==1.2.0
django-reset==0.2.0
django-reversion==1.8.1
django-stdimage==0.5.10
django-userena==1.3.0
django-xadmin==0.5.0
easy-thumbnails==2.0.1
html2text==2014.4.5
ipython==2.1.0
phonenumbers==6.1.0
simplejson==3.5.2
six==1.7.2
sqlparse==0.1.11
wsgiref==0.1.2
xlwt==0.7.5
